+++ src/test/java/com/antibdd/runner/RunCucumberTest.java
@@ -1,0 +1,10 @@
+package com.antibdd.runner;
+
+import cucumber.api.CucumberOptions;
+import cucumber.api.junit.Cucumber;
+import org.junit.runner.RunWith;
+
+@RunWith(Cucumber.class)
+@CucumberOptions(plugin = {"pretty"})
+public class RunCucumberTest {
+}
