--- src/test/java/com/kazimir/peter/RunCucumberTest.java
@@ -1,10 +1,0 @@
-package com.kazimir.peter;
-
-import cucumber.api.CucumberOptions;
-import cucumber.api.junit.Cucumber;
-import org.junit.runner.RunWith;
-
-@RunWith(Cucumber.class)
-@CucumberOptions(plugin = {"pretty"})
-public class RunCucumberTest {
-}
